+++ app/src/main/java/com/yterletskyi/happyfriend/features/pin/domain/PinButtonModel.kt
@@ -1,0 +1,6 @@
+package com.yterletskyi.happyfriend.features.pin.domain
+
+sealed interface PinButtonModel {
+    data class Number(val value: Int) : PinButtonModel
+    object Erase : PinButtonModel
+}
