+++ app/src/main/java/com/yterletskyi/happyfriend/features/pin/ui/PinKeyboardButtonView.kt
@@ -1,4 +1,4 @@
-package com.yterletskyi.happyfriend.common.view
+package com.yterletskyi.happyfriend.features.pin.ui
 
 import android.content.Context
 import android.content.res.ColorStateList
