+++ app/src/main/res/layout/fragment_pin.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".features.pin.ui.PinFragment">
 
-    <com.yterletskyi.happyfriend.common.view.PinProgressView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinProgressView
         android:id="@+id/pinProgressView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -15,10 +15,9 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:progressCircleColor="?attr/colorPrimary"
-        app:steps="4" />
+        app:progressCircleColor="?attr/colorPrimary" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardView
         android:id="@+id/pinKeyboardView"
         android:layout_width="0dp"
         android:layout_height="0dp"
