+++ app/src/main/res/layout/view_pin_keyboard.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content"
     tools:parentTag="androidx.constraintlayout.widget.ConstraintLayout">
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn1"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_one"
@@ -15,7 +15,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn2"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_two"
@@ -24,7 +24,7 @@
         app:layout_constraintStart_toEndOf="@id/btn1"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn3"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_three"
@@ -33,7 +33,7 @@
         app:layout_constraintStart_toEndOf="@id/btn2"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn4"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_four"
@@ -42,7 +42,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/btn1" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn5"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_five"
@@ -51,7 +51,7 @@
         app:layout_constraintStart_toEndOf="@id/btn4"
         app:layout_constraintTop_toBottomOf="@id/btn2" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn6"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_six"
@@ -60,7 +60,7 @@
         app:layout_constraintStart_toEndOf="@id/btn5"
         app:layout_constraintTop_toBottomOf="@id/btn3" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn7"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_seven"
@@ -69,7 +69,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/btn4" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn8"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_eight"
@@ -78,7 +78,7 @@
         app:layout_constraintStart_toEndOf="@id/btn7"
         app:layout_constraintTop_toBottomOf="@id/btn5" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn9"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_nine"
@@ -97,7 +97,7 @@
         app:layout_constraintTop_toBottomOf="@id/btn7"
         tools:background="#000" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btn0"
         style="@style/PinKeyboardButtonStyle"
         android:text="@string/pin_button_zero"
@@ -106,7 +106,7 @@
         app:layout_constraintStart_toStartOf="@id/btn8"
         app:layout_constraintTop_toBottomOf="@id/btn7" />
 
-    <com.yterletskyi.happyfriend.common.view.PinKeyboardButtonView
+    <com.yterletskyi.happyfriend.features.pin.ui.PinKeyboardButtonView
         android:id="@+id/btnErase"
         style="@style/PinKeyboardButtonStyle"
         android:src="@drawable/ic_round_backspace_24"
